+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: jballest <jballest@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 17:14:29 by jballest          #+#    #+#             */
-/*   Updated: 2021/04/01 15:01:50 by jballest         ###   ########.fr       */
+/*   Updated: 2021/04/01 16:06:52 by jballest         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ int				ft_isdigit(int c);
 int				ft_isalnum(int c);
 int				ft_isascii(int c);
 int				ft_isprint(int c);
-int				ft_isnegative(long long n);
+int				ft_isnegative(int n);
 int				ft_toupper(int c);
 int				ft_tolower(int c);
 int				ft_atoi(const char *str);
